Refactor API code
